CA-285401 add module Readln
This module helps reading lines from a file descriptor. Unlike
read_line(), it can read multiple lines when they are available. This is
a problem when reading lines triggered by select() or epoll(). It is
desirable to read all available lines rather than just the first, as
read_line() does.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>